refactor: 변경된 port 적용
