+++ src/main/java/com/mobble/mobbleserver/application/notification/core/service/NotificationModifyService.java
@@ -1,8 +1,10 @@
 package com.mobble.mobbleserver.application.notification.core.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
+import com.mobble.mobbleserver.application.notification.core.port.provided.NotificationIssuePort;
+import com.mobble.mobbleserver.application.notification.core.port.provided.NotificationMarkPort;
 import com.mobble.mobbleserver.domain.member.Member;
-import com.mobble.mobbleserver.infrastructure.persistence.member.JpaMemberRepository;
 import com.mobble.mobbleserver.domain.notification.core.Notification;
 import com.mobble.mobbleserver.infrastructure.persistence.notification.core.JpaNotificationRepository;
 import com.mobble.mobbleserver.domain.notification.device.DeviceToken;
@@ -13,7 +15,6 @@ import com.mobble.mobbleserver.domain.notification.setting.NotificationSetting;
 import com.mobble.mobbleserver.infrastructure.persistence.notification.setting.JpaNotificationSettingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,11 +26,12 @@ import java.util.Objects;
 import static com.mobble.mobbleserver.infrastructure.web.notification.dto.NotificationDto.ReadAllReq;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class NotificationService {
+public class NotificationModifyService implements NotificationIssuePort, NotificationMarkPort {
 
-    private final JpaMemberRepository memberRepository;
+    private final MemberReadPort memberReadPort;
+
     private final JpaNotificationRepository jpaNotificationRepository;
     private final JpaNotificationSettingRepository settingRepository;
     private final JpaDeviceTokenRepository jpaDeviceTokenRepository;
@@ -38,9 +40,9 @@ public class NotificationService {
     private final ObjectMapper objectMapper;
     private final ApplicationEventPublisher events;
 
-    @Transactional
+    @Override
     public Long issue(SendNotificationCommand command) {
-        Member receiver = memberRepository.getReferenceById(command.receiverId());
+        Member receiver = memberReadPort.findByIdAndIsDeletedFalse(command.receiverId()).orElseThrow();
 
         Notification saved = jpaNotificationRepository.save(
                 Notification.create(receiver, command.type(), command.title(), command.content(), command.targetType(), command.targetId())
@@ -70,11 +72,7 @@ public class NotificationService {
         return saved.getId();
     }
 
-    public List<Notification> getList(Long memberId, Long cursorId, int size) {
-        return jpaNotificationRepository.findSlice(memberId, cursorId, PageRequest.of(0, size));
-    }
-
-    @Transactional
+    @Override
     public void markRead(Long memberId, Long notificationId) {
         Notification n = jpaNotificationRepository.findById(notificationId)
                 .orElseThrow(() -> new NoSuchElementException("notification not found"));
@@ -82,15 +80,10 @@ public class NotificationService {
         n.markAsRead();
     }
 
-    @Transactional
+    @Override
     public void markAllRead(Long memberId, ReadAllReq request) {
         Long upToId = request == null ? null : request.upToId();
         jpaNotificationRepository.markAllRead(memberId, upToId);
-    }
-
-    @Transactional(readOnly = true)
-    public long unreadCount(Long memberId) {
-        return jpaNotificationRepository.countByReceiver_IdAndIsReadFalse(memberId);
     }
 
     public record PushReadyEvent() {
